+++ app/states/play-states/fight-state.js
@@ -1,10 +1,7 @@
-let debug = {
-    log: require('debug')('7balls:fight-state:log')
-};
-
 const FIGHT = require('../../constants/fight');
 
 import ArtificialIntelligence from '../../helpers/artificial-intelligence';
+import Enemy from '../../models/enemy';
 
 let assert = require('assert');
 let isString = require('lodash.isstring');
@@ -12,45 +9,38 @@ let utils = require('../../helpers/utils');
 let { displaySingleLineMessage } = require('../../helpers/message');
 let OptionsEnemyMixin = require('../mixins/options-enemy-mixin');
 let OptionsPlayerMixin = require('../mixins/options-player-mixin');
-import Enemy from '../../models/enemy';
 
-function defineAnimations(character, reduceByHalf, revertDefaultSize) {
-    let sprite = character.getSprite();
+function defineAnimations(character, revertDefaultSize) {
+    let $sprite = character.getSprite();
     let width = 150;
     let height = 200;
 
     function resizeMaximum() {
-        sprite.body.setSize(width, height, 0, 0);
+        $sprite.body.setSize(width, height, 0, 0);
     }
 
-    let sitting = sprite.animations.add('sitting', [4, 5], 4, false);
+    $sprite.animations.add('standing', [0, 1, 2], 4, true);
+    $sprite.animations.add('win', [3, 4], 4, false);
+    $sprite.animations.add('died', [6, 7], 4, false);
 
-    sitting.onStart.add(reduceByHalf);
-    sitting.onComplete.add(revertDefaultSize);
-
-    let kicking = sprite.animations.add('kicking', [12, 13], 16, false);
-
+    let kicking = $sprite.animations.add('kicking', [9, 10], 16, false);
     kicking.onStart.add(resizeMaximum);
     kicking.onComplete.add(revertDefaultSize);
+    kicking.onComplete.add(() => {
+        $sprite.play('standing');
+    });
 
-    let boxing = sprite.animations.add('boxing', [16, 17], 16, false);
-
+    let boxing = $sprite.animations.add('boxing', [12, 13], 16, false);
     boxing.onStart.add(resizeMaximum);
     boxing.onComplete.add(revertDefaultSize);
-
-    sprite.animations.add('standing', [0, 1, 2], 4, true);
-    sprite.animations.add('jumping', [8, 9], 4, false);
-    sprite.animations.add('win', [20, 21], 4, false);
-    sprite.animations.add('died', [24, 25], 4, false);
-
-    sprite.play('standing');
-
-    debug.log('Character "%s" is STANDING', character.title);
+    boxing.onComplete.add(() => {
+        $sprite.play('standing');
+    });
 }
 
-function resetCharacterVelocity(sprite) {
-    sprite.body.velocity.x = 0;
-    sprite.body.velocity.y += FIGHT.FALL_SPEED;
+function resetCharacterVelocity($sprite) {
+    $sprite.body.velocity.x = 0;
+    $sprite.body.velocity.y += FIGHT.FALL_SPEED;
 }
 
 export default class FightState extends Phaser.State {
@@ -67,11 +57,6 @@ export default class FightState extends Phaser.State {
         }
     };
 
-    init() {
-        Object.assign(this, OptionsEnemyMixin);
-        Object.assign(this, OptionsPlayerMixin);
-    }
-
     displayLogo() {
         this.add.image((this.game.width / 2) - (this.cache.getImage('logo').width / 2), 5, 'logo');
     }
@@ -83,13 +68,13 @@ export default class FightState extends Phaser.State {
     }
 
     _defineDefaultProperties(character) {
-        let sprite = character.getSprite();
+        let $sprite = character.getSprite();
 
         // Tutaj tworzymy ciało dla naszej postaci.
-        this.physics.arcade.enable(sprite);
+        this.physics.arcade.enable($sprite);
 
-        sprite.body.bounce.setTo(0, 0.1);
-        sprite.body.collideWorldBounds = true;
+        $sprite.body.bounce.setTo(0, 0.1);
+        $sprite.body.collideWorldBounds = true;
     }
 
     /**
@@ -97,50 +82,32 @@ export default class FightState extends Phaser.State {
      * @param {string} orientation [left, right]
      * @access protected
      */
-    _setupOrientation(character, orientation) {
-        let sprite = character.getSprite();
+    _getOrientationResetSize(character, orientation) {
+        let $sprite = character.getSprite();
         let width = 150;
         let height = 200;
-        let defaultSize = null;
-        let reduceByHalf = null;
 
-        if (orientation === sprite.orientation) {
+        if (orientation === $sprite.orientation) {
             // Nie zmieniamy orientacji na już wybraną.
             return;
         }
 
-        sprite.orientation = orientation;
+        $sprite.orientation = orientation;
 
-        switch (orientation) {
-            case 'left':
-                defaultSize = () => {
-                    sprite.body.setSize((width * 2) / 3, height, 0, 0);
-                };
+        return () => {
+            switch (orientation) {
+                case 'left':
+                    $sprite.body.setSize((width * 2) / 3, height, 0, 0);
+                    break;
 
-                reduceByHalf = () => {
-                    sprite.body.setSize((width * 2) / 3, height / 2, 0, height / 2);
-                };
+                case 'right':
+                    $sprite.body.setSize((width * 2) / 3, height, width / 3, 0);
+                    break;
 
-                defineAnimations(character, reduceByHalf, defaultSize);
-                defaultSize();
-                break;
-
-            case 'right':
-                defaultSize = () => {
-                    sprite.body.setSize((width * 2) / 3, height, width / 3, 0);
-                };
-
-                reduceByHalf = () => {
-                    sprite.body.setSize((width * 2) / 3, height / 2, width / 3, height / 2);
-                };
-
-                defineAnimations(character, reduceByHalf, defaultSize);
-                defaultSize();
-                break;
-
-            default:
-            // no default
-        }
+                default:
+                // no default
+            }
+        };
     }
 
     _addAvatar(x, y, key) {
@@ -161,63 +128,47 @@ export default class FightState extends Phaser.State {
     }
 
     _setupSprite(x, y, character, anchor = [0, 1]) {
-        character.createSprite(this, x, y, `${character.id}-spritesheet`);
+        character.createSprite(this, x, y, `${character.id}-fight`);
 
-        let sprite = character.getSprite();
-        sprite.anchor.setTo(...anchor);
+        let $sprite = character.getSprite();
+        $sprite.anchor.setTo(...anchor);
 
         this._defineDefaultProperties(character);
-
         this._setupMoves(character);
     }
 
     _setupMoves(character) {
-        let sprite = character.getSprite();
+        let $sprite = character.getSprite();
 
-        sprite.events.onLeft = new Phaser.Signal();
-        sprite.events.onLeft.add(() => {
-            sprite.body.velocity.x -= FIGHT.HORIZONTAL_SPEED;
-
-            debug.log('Character "%s" is LEFT', character.title);
+        $sprite.events.onLeft = new Phaser.Signal();
+        $sprite.events.onLeft.add(() => {
+            $sprite.body.velocity.x -= FIGHT.HORIZONTAL_SPEED;
         });
 
-        sprite.events.onRight = new Phaser.Signal();
-        sprite.events.onRight.add(() => {
-            sprite.body.velocity.x += FIGHT.HORIZONTAL_SPEED;
-
-            debug.log('Character "%s" is RIGHT', character.title);
+        $sprite.events.onRight = new Phaser.Signal();
+        $sprite.events.onRight.add(() => {
+            $sprite.body.velocity.x += FIGHT.HORIZONTAL_SPEED;
         });
 
-        sprite.events.onSitting = new Phaser.Signal();
-        sprite.events.onSitting.add(() => {
-            sprite.play('sitting');
-
-            debug.log('Character "%s" is SITTING', character.title);
-        });
-
-        sprite.events.onJumping = new Phaser.Signal();
-        sprite.events.onJumping.add(() => {
-            if (!sprite.body.onFloor()) {
+        $sprite.events.onJumping = new Phaser.Signal();
+        $sprite.events.onJumping.add(() => {
+            if (!$sprite.body.onFloor()) {
                 return;
             }
 
-            sprite.body.velocity.y -= FIGHT.JUMP;
-            sprite.play('jumping');
-            debug.log('Character "%s" is JUMPING', character.title);
+            $sprite.body.velocity.y -= FIGHT.JUMP;
         });
 
-        sprite.events.onKicking = new Phaser.Signal();
-        sprite.events.onKicking.add(() => {
-            sprite.play('kicking');
-            debug.log('Character "%s" is KICKING', character.title);
+        $sprite.events.onKicking = new Phaser.Signal();
+        $sprite.events.onKicking.add(() => {
+            $sprite.play('kicking');
         });
 
-        sprite.events.onBoxing = new Phaser.Signal();
-        sprite.events.onBoxing.add(() => {
-            sprite.play('boxing');
-            debug.log('Character "%s" is BOXING', character.title);
+        $sprite.events.onBoxing = new Phaser.Signal();
+        $sprite.events.onBoxing.add(() => {
+            $sprite.play('boxing');
         });
-        sprite.events.onDied = new Phaser.Signal();
+        $sprite.events.onDied = new Phaser.Signal();
     }
 
     _updateOptionsLvL(character) {
@@ -225,7 +176,7 @@ export default class FightState extends Phaser.State {
     }
 
     _setupKeyboard() {
-        let playerSprite = this.game.player.getSprite();
+        let $playerSprite = this.game.player.getSprite();
 
         let c = this.input.keyboard.addKey(Phaser.Keyboard.C);
         let x = this.input.keyboard.addKey(Phaser.Keyboard.X);
@@ -240,10 +191,10 @@ export default class FightState extends Phaser.State {
             Phaser.Keyboard.SPACEBAR
         ]);
 
-        c.onDown.add(() => playerSprite.events.onKicking.dispatch());
-        x.onDown.add(() => playerSprite.events.onBoxing.dispatch());
-        up.onDown.add(() => playerSprite.events.onJumping.dispatch());
-        space.onDown.add(() => playerSprite.events.onJumping.dispatch());
+        c.onDown.add(() => $playerSprite.events.onKicking.dispatch());
+        x.onDown.add(() => $playerSprite.events.onBoxing.dispatch());
+        up.onDown.add(() => $playerSprite.events.onJumping.dispatch());
+        space.onDown.add(() => $playerSprite.events.onJumping.dispatch());
     }
 
     _setupEnemy() {
@@ -255,17 +206,17 @@ export default class FightState extends Phaser.State {
     }
 
     create() {
+        Object.assign(this, OptionsEnemyMixin, OptionsPlayerMixin);
+
         this.add.image(0, 0, 'bg-select-player');
 
         this._setupWorld();
-
         this._setupEnemy();
 
         this._setupSprite(150, 360, this.game.player);
         this._setupSprite(650, 360, this.game.enemy, [1, 1]);
 
-        this._setupOrientation(this.game.player, 'left');
-        this._setupOrientation(this.game.enemy, 'right');
+        this._setupFightersActions();
 
         this._setupPlayerOptions();
         this._setupEnemyOptions();
@@ -275,6 +226,18 @@ export default class FightState extends Phaser.State {
 
         this.displayLogo();
         displaySingleLineMessage(this.game, `${this.game.locale.FIGHT_STATE_WELCOME}`);
+    }
+
+    _setupFightersActions() {
+        let playerResetSize = this._getOrientationResetSize(this.game.player, 'left');
+        defineAnimations(this.game.player, playerResetSize);
+        playerResetSize();
+        this.game.player.getSprite().play('standing');
+
+        let enemyResetSize = this._getOrientationResetSize(this.game.enemy, 'right');
+        defineAnimations(this.game.enemy, enemyResetSize);
+        enemyResetSize();
+        this.game.enemy.getSprite().play('standing');
     }
 
     update() {
@@ -288,11 +251,11 @@ export default class FightState extends Phaser.State {
 
     _setupFight() {
         let context = this;
-        let playerSprite = this.game.player.getSprite();
-        let enemySprite = this.game.enemy.getSprite();
+        let $playerSprite = this.game.player.getSprite();
+        let $enemySprite = this.game.enemy.getSprite();
 
         function isCollision() {
-            return context.physics.arcade.overlap(enemySprite, playerSprite);
+            return context.physics.arcade.overlap($enemySprite, $playerSprite);
         }
 
         function handlePlayerBlow(points) {
@@ -302,9 +265,9 @@ export default class FightState extends Phaser.State {
             }
         }
 
-        playerSprite.events.onKicking.add(() => handlePlayerBlow(FIGHT.KICKING_POINTS));
-        playerSprite.events.onBoxing.add(() => handlePlayerBlow(FIGHT.BOXING_POINTS));
-        playerSprite.events.onDied.add(() => this._finishFight('died', 'win'));
+        $playerSprite.events.onKicking.add(() => handlePlayerBlow(FIGHT.KICKING_POINTS));
+        $playerSprite.events.onBoxing.add(() => handlePlayerBlow(FIGHT.BOXING_POINTS));
+        $playerSprite.events.onDied.add(() => this._finishFight('died', 'win'));
 
         function handleEnemyBlow(points) {
             if (isCollision()) {
@@ -313,16 +276,16 @@ export default class FightState extends Phaser.State {
             }
         }
 
-        enemySprite.events.onKicking.add(() => handleEnemyBlow(FIGHT.KICKING_POINTS));
-        enemySprite.events.onBoxing.add(() => handleEnemyBlow(FIGHT.BOXING_POINTS));
-        enemySprite.events.onDied.add(() => this._finishFight('win', 'died'));
+        $enemySprite.events.onKicking.add(() => handleEnemyBlow(FIGHT.KICKING_POINTS));
+        $enemySprite.events.onBoxing.add(() => handleEnemyBlow(FIGHT.BOXING_POINTS));
+        $enemySprite.events.onDied.add(() => this._finishFight('win', 'died'));
 
-        ArtificialIntelligence.setup(this, enemySprite);
+        ArtificialIntelligence.setup(this, $enemySprite);
     }
 
-    _getLocaleStatus(playerSate) {
-        assert(isString(playerSate));
-        let status = playerSate.toLowerCase();
+    _getLocaleStatus(playerState) {
+        assert(isString(playerState));
+        let status = playerState.toLowerCase();
 
         let strategies = new Map();
         strategies.set('win', this.game.locale.FIGHT_STATE_PLAYER_WIN);
@@ -334,35 +297,35 @@ export default class FightState extends Phaser.State {
         return locale;
     }
 
-    _finishFight(playerSate, enemyState) {
+    _finishFight(playerState, enemyState) {
         let player = this.game.player;
-        let playerSprite = player.getSprite();
+        let $playerSprite = player.getSprite();
         let enemy = this.game.enemy;
-        let enemySprite = enemy.getSprite();
+        let $enemySprite = enemy.getSprite();
 
-        displaySingleLineMessage(this.game, `${player.title} ${this._getLocaleStatus(playerSate)}`);
+        displaySingleLineMessage(this.game, `${player.title} ${this._getLocaleStatus(playerState)}`);
 
         // Wyłączamy wsparcie klawiatury w grze.
         this.input.keyboard.enabled = false;
 
         // Sprowadzamy postacie na ziemie.
-        playerSprite.y = 0;
-        enemySprite.y = 0;
+        $playerSprite.body.velocity.setTo(0, 0);
+        $enemySprite.body.velocity.setTo(0, 0);
 
         // Włączamy animacje: zwycięstwa i porażki.
-        playerSprite.play(playerSate);
-        enemySprite.play(enemyState);
+        $playerSprite.play(playerState);
+        $enemySprite.play(enemyState);
 
         // Uśmiercam obu, aby żadne animacje nie były więcej wykonywane.
         // Nie używamy .kill() bo wtedy nie wykona się ost. animacja (zwycięstwa i porażki).
-        playerSprite.alive = false;
-        enemySprite.alive = false;
+        $playerSprite.alive = false;
+        $enemySprite.alive = false;
 
         utils.timeout(this, Phaser.Timer.SECOND * 2, () => {
             // Przywracamy wsparcie klawiatury w grze.
             this.input.keyboard.enabled = true;
 
-            if (playerSate === 'died') {
+            if (playerState === 'died') {
                 this.game.emit('game:over', { enemy: this.game.enemy });
             } else {
                 // Usuwamy pierwszego, pokonanego wroga.
@@ -373,27 +336,23 @@ export default class FightState extends Phaser.State {
     }
 
     _handleKeyboard() {
-        let playerSprite = this.game.player.getSprite();
+        let $playerSprite = this.game.player.getSprite();
         let keyboard = this.input.keyboard;
 
         if (keyboard.isDown(Phaser.Keyboard.LEFT)) {
-            playerSprite.events.onLeft.dispatch();
+            $playerSprite.events.onLeft.dispatch();
         } else if (keyboard.isDown(Phaser.Keyboard.RIGHT)) {
-            playerSprite.events.onRight.dispatch();
-        }
-
-        if (keyboard.isDown(Phaser.Keyboard.DOWN)) {
-            playerSprite.events.onSitting.dispatch();
+            $playerSprite.events.onRight.dispatch();
         }
     }
 
     render() {
-        let playerSprite = this.game.player.getSprite();
-        // this.game.debug.bodyInfo(playerSprite, 25, 25);
-        this.game.debug.body(playerSprite);
+        let $playerSprite = this.game.player.getSprite();
+        // this.game.debug.bodyInfo($playerSprite, 25, 25);
+        this.game.debug.body($playerSprite);
 
-        let enemySprite = this.game.enemy.getSprite();
-        // this.game.debug.bodyInfo(enemySprite, 25, 225);
-        this.game.debug.body(enemySprite);
+        let $enemySprite = this.game.enemy.getSprite();
+        // this.game.debug.bodyInfo($enemySprite, 25, 225);
+        this.game.debug.body($enemySprite);
     }
 }